# Conflicts:
#	app/src/main/java/org/akvo/flow/activity/SurveyActivity.java
#	app/src/main/res/layout/drawer_header_layout.xml

+++ app/src/main/java/org/akvo/flow/activity/SurveyActivity.java
@@ -251,8 +251,10 @@ public class SurveyActivity extends AppCompatActivity implements RecordListListe
         } else {
             activityJustCreated = false;
             // Delete empty responses, if any
-            mDatabase.deleteEmptySurveyInstances();
-            mDatabase.deleteEmptyRecords();
+            if (mDatabase != null) {
+                mDatabase.deleteEmptySurveyInstances();
+                mDatabase.deleteEmptyRecords();
+            }
 
             ViewApkData apkData = apkUpdateStore.getApkData();
             boolean shouldNotifyUpdate = apkUpdateStore.shouldNotifyNewVersion();
@@ -423,7 +425,8 @@ public class SurveyActivity extends AppCompatActivity implements RecordListListe
     }
 
     private void displayForm(String surveyedLocaleId, User user) {
-        Survey registrationForm = mDatabase.getRegistrationForm(mSurveyGroup);
+        Survey registrationForm =
+                mDatabase != null ? mDatabase.getRegistrationForm(mSurveyGroup) : null;
         if (registrationForm == null) {
             Toast.makeText(this, R.string.error_missing_form, Toast.LENGTH_LONG).show();
             return;
@@ -497,8 +500,10 @@ public class SurveyActivity extends AppCompatActivity implements RecordListListe
 
     @OnClick(R.id.add_data_point_fab)
     void onAddDataPointTap() {
-        addDataPointFab.setEnabled(false);
-        String newLocaleId = mDatabase.createSurveyedLocale(mSurveyGroup.getId());
-        onRecordSelected(newLocaleId);
+        if (mDatabase != null) {
+            addDataPointFab.setEnabled(false);
+            String newLocaleId = mDatabase.createSurveyedLocale(mSurveyGroup.getId());
+            onRecordSelected(newLocaleId);
+        }
     }
 }
